requirements.txt
Django==1.7.2
dj-database-url==0.3.0
dj-static==0.0.6
django-tastypie==0.12.1
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
python-bitcoinlib==0.2.2-SNAPSHOT
python-dateutil==2.4.0
python-mimeparse==0.1.4
six==1.9.0
static3==0.5.1
virtualenv==12.0.5
wsgiref==0.1.2
